Reexport the app in embed
